refactor: move OrderProgressBar to module
