+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/container/ProgressImageWrapper.tsx
