+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/container/RenderProgressTopSection.tsx
