+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/container/StepComponent.tsx
