+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/container/TopSections.tsx
