+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/container/styled.ts
