+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/helpers.ts
