+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/sharedStyled.ts
