+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/steps/AnimatedToken.tsx
