+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/steps/InitialStep.tsx
