+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/steps/SolvingStep.tsx
