+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/styled.ts
